:building_construction: #148 Installation page : split unlocked / locked in different folders
