+++ client/src/pages/Installation/unlocked/installation-unlocked.scss
@@ -1,34 +1,8 @@
-@import '../../assets/styles/global';
+@import '../../../assets/styles/global';
 
 .page-installation {
   background-color: $color-blue;
   height: 100vh;
-
-  &--locked {
-    &__presentation {
-      &__title {
-        display: inline-block;
-        margin: 3.5rem 0;
-        width: 100%;
-        color: $color-green;
-        font-size: 4rem;
-        text-align: center;
-        font-family: $font-infini-regular;
-        text-transform: uppercase;
-      }
-
-      &__text-content {
-        text-align: center;
-        max-width: $content-width;
-        padding: 3rem $content-padding 0 $content-padding;
-        margin: auto;
-        font-family: $font-infini-regular;
-        color: $color-pink;
-        font-size: 1.5rem;
-        line-height: 1.3em;
-      }
-    }
-  }
 
   &__wrapper {
     height: 100%;
